Enable the CLI (for local work)
Problem
-------

We'd like a CLI so you could test the underlying image functionality
without always going through a bot.

Solution
--------

Properly configure elixir (in `mix.exs`) to allow for a CLI.

The slight issue with this is that if you
```
mix build.escript
./slack_color_theme_generator image.png
```
while the process is running, the SlackBot is also turned on.

Ideally we'd have them as separate apps.

For now this gets us a step forward so we could have the app run as a
CLI.  Sadly it still needs a SLACK_API_TOKEN to run.

step 1...