Add the ability to run the app without a SLACK_API_TOKEN
problem
-------

Without SLACK_API_TOKEN, the App doesn't run because the
`Slack.Bot` initialization fails.  This means you can't run
tests or the commandline app without the token.

That's dumb.

with @gabriellet

solution
--------

Move things into GenServers that could conditionally start
themselves.

Add ImageProcessor Module with Server and Impl.
The impl provides the `compute_theme` method.  The server handles
messaging to and from other consumers.  Read the book.

Move `SlackBot` to `SlackBotWrapper` and `SlackColorThemeGenerator`
functions to the `ImageProcessor.Impl`.

Add `ImageProcessor` to the Supervisor tree and get CLI and Slack.Bot
to make calls to `ImageProcessor` run through the GenServer.

Added `SlackBotWrapper` which manages starting (or not) the
`Slack.Bot` worker if the token is available (or not).

Notes:
------

Followed examples in from Programming Elixir 1.6 Chapter 17: